Renamed `data_interface` to `data_requesting`
 - Added `specific_coverage.sh`, to only run specific tests.
 - Tweaks to response handling in Requester.
