+++ Tests/specific_coverage.sh
@@ -1,0 +1,4 @@
+coverage run -s . -p 'TestRequesting.py' -m unittest discover
+coverage report -m
+coverage html
+exit
