+++ core/data_fetching/DataLoader.py
@@ -8,7 +8,7 @@ from core.wubrg import COLOR_COMBINATIONS
 from core.utilities.auto_logging import logging
 from core.utilities import save_json_file, load_json_file
 
-from core.data_interface.Requester import Requester
+from core.data_requesting.Requester import Requester
 from core.game_metadata import CardManager
 from core.data_fetching.utils.settings import DATA_DIR_LOC, DATA_DIR_NAME
