+++ core/data_requesting/Request17Lands.py
@@ -7,8 +7,8 @@ from datetime import date, datetime
 import json
 import pandas as pd
 
-from core.data_interface.utils import *
-from core.data_interface.Requester import Requester
+from core.data_requesting.utils import *
+from core.data_requesting.Requester import Requester
 
 
 # Adapted from 'https://github.com/diogojapinto/mtg-data-mining/blob/main/utils/api_clients/seventeen_lands/client.py'
