+++ core/data_requesting/utils/__init__.py
@@ -2,9 +2,9 @@
 This package helps simplify managing code which requests data from websites.
 """
 
-from core.data_interface.utils.consts import *
-from core.data_interface.utils.funcs import *
-from core.data_interface.utils.settings import *
+from core.data_requesting.utils.consts import *
+from core.data_requesting.utils.funcs import *
+from core.data_requesting.utils.settings import *
 
 from_consts = ['BASE_17L_URL', 'COLOR_17L_URL', 'EXPANSIONS_17L_URL', 'FORMATS_17L_URL', 'PLAY_DRAW_17L_URL',
                'COLOR_RATING_17L_URL', 'CARD_RATING_17L_URL', 'CARD_EVAL_17L_URL', 'TROPHY_17L_URL',
