+++ core/data_requesting/utils/funcs.py
